Added Old Mod Copy
Added a separate folder containing a copy of the old mod. Also added a new folder where I will store my personalized version of the code, used to tweak the game to how I prefer it even if that isn't compatible with the base mod.

Currently it's just a copy because I already made some changes and I want to be able to compare my copy to the original mod.